Track down issues w/ starting Components.
The main problem stemmed from Components that were returning nil and
documentation/error messages that didn't make it blatantly obvious that
this is actually a major problem.

Up to the point of failing because the things I was doing in my
background threads don't mesh well with the way I'm trying to do those
same things (or not) now.