I think I may have hit on the magical combination.
The ns declaration seems to fail unless the import happens after the ns
involved in the import has already been required. It also seems as
though that ns needs to do a gen-class so its output can be seen.

I could still be totally out in left field, but I think those changes
are what got me past some basic linking errors. Which is a truly
uncomfortable state of affairs.