Try using clojurewerkz' buffy library
Initially just for transmitting the port where the reader is
listening. I suspect this particular use case is totally invalid. But
I *do* need a way to transmit binary data. And something like protocol
buffers feels like overkill.

Or I may have the entire idea totally upside down, and that's what I
should be using.

Start with this. At least it's *a* step, even if it isn't precisely forward.